requirements.txt
chromadb>=0.5.23
httpx>=0.28.1
openai>=1.70.0
pypdf>=5.4.0
python-docx>=1.1.2
python-dotenv>=1.0.1
python-pptx>=1.0.2
streamlit>=1.43.2

